Refactor with inheritance
