+++ src/Column/BigIntColumnSchema.php
@@ -4,36 +4,22 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql\Column;
 
-use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Schema\SchemaInterface;
+use Yiisoft\Db\Schema\Column\BigIntColumnSchema as BaseBigIntColumnSchema;
 
-final class BigIntColumnSchema extends AbstractIntegerColumnSchema
+final class BigIntColumnSchema extends BaseBigIntColumnSchema implements IntegerColumnSchemaInterface
 {
-    public function __construct(string $name)
-    {
-        parent::__construct($name);
+    /**
+     * @var string|null Name of an associated sequence if column is auto incremental.
+     */
+    private string|null $sequenceName = null;
 
-        $this->type(SchemaInterface::TYPE_BIGINT);
+    public function getSequenceName(): string|null
+    {
+        return $this->sequenceName;
     }
 
-    public function dbTypecast(mixed $value): int|string|ExpressionInterface|null
+    public function sequenceName(string|null $sequenceName): void
     {
-        return match (true) {
-            is_int($value), $value === null, $value instanceof ExpressionInterface => $value,
-            $value === '' => null,
-            $value === false => 0,
-            PHP_INT_MIN <= $value && $value <= PHP_INT_MAX => (int) $value,
-            default => (string) $value,
-        };
-    }
-
-    public function phpTypecast(mixed $value): int|string|null
-    {
-        /** @psalm-var int|string|null $value */
-        return match (true) {
-            $value === null => null,
-            PHP_INT_MIN <= $value && $value <= PHP_INT_MAX => (int) $value,
-            default => (string) $value,
-        };
+        $this->sequenceName = $sequenceName;
     }
 }
