+++ src/Column/BooleanColumnSchema.php
@@ -4,30 +4,10 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Pgsql\Column;
 
-use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Schema\Column\AbstractColumnSchema;
-use Yiisoft\Db\Schema\SchemaInterface;
+use Yiisoft\Db\Schema\Column\BooleanColumnSchema as BaseBooleanColumnSchema;
 
-final class BooleanColumnSchema extends AbstractColumnSchema
+final class BooleanColumnSchema extends BaseBooleanColumnSchema
 {
-    public function __construct(string $name)
-    {
-        parent::__construct($name);
-
-        $this->type(SchemaInterface::TYPE_BOOLEAN);
-        $this->phpType(SchemaInterface::PHP_TYPE_BOOLEAN);
-    }
-
-    public function dbTypecast(mixed $value): bool|ExpressionInterface|null
-    {
-        /** Optimized for performance. Do not merge cases or change order. */
-        return match (true) {
-            $value, $value === false, $value === null, $value instanceof ExpressionInterface => $value,
-            $value === '' => null,
-            default => (bool) $value,
-        };
-    }
-
     public function phpTypecast(mixed $value): bool|null
     {
         if ($value === null) {
